fix(gsdump): validate GS dump inputs and audit resolvers
Adds validation across the dump reader, audit resolvers, pixel diff, and the GIF
interpreter's alpha path, and lets a bare probe output name resolve against the
current directory.

Known issue: GsDumpAuditTests.TextureDump_LabelsFramebufferFeedbackSource now
sees source "rt_cache" where it expects "framebuffer"; whether the render-target
cache should mask the framebuffer-feedback label is unresolved.

Work produced in the parallel Codex session; committed here after verifying
both targets build and the suite runs (see the batch report for known failures).

+++ src/NeversoftMultitool/Core/Formats/GsDump/GsDumpFile.cs
@@ -94,7 +94,7 @@ internal sealed class GsDumpFile
 
             if (screenshotWidth > 0 &&
                 screenshotHeight > 0 &&
-                screenshotSize == screenshotWidth * screenshotHeight * 4 &&
+                screenshotSize == (long)screenshotWidth * screenshotHeight * 4 &&
                 screenshotOffset >= 0 &&
                 screenshotOffset <= headerBlock.Length &&
                 screenshotSize <= headerBlock.Length - screenshotOffset)
